+++ lib/driver.js
@@ -14,8 +14,7 @@ import commands from './commands/index';
 import { detectUdid, getAndCheckXcodeVersion, getAndCheckIosSdkVersion,
          adjustWDAAttachmentsPermissions, checkAppPresent, getDriverInfo,
          clearSystemFiles, translateDeviceName, normalizeCommandTimeouts,
-         DEFAULT_TIMEOUT_KEY } from './utils';
-import { killAppUsingAppName } from './wda/utils';
+         DEFAULT_TIMEOUT_KEY, killAppUsingAppName } from './utils';
 import { getConnectedDevices, runRealDeviceReset, installToRealDevice,
          getRealDeviceObj } from './real-device-management';
 import { NoSessionProxy } from "./wda/no-session-proxy";
