+++ lib/simulator-management.js
@@ -1,7 +1,7 @@
 import path from 'path';
 import { getSimulator } from 'appium-ios-simulator';
 import { createDevice, getDevices, terminate } from 'node-simctl';
-import { killAppUsingAppName } from './wda/utils';
+import { killAppUsingAppName } from './utils';
 import _ from 'lodash';
 import log from './logger';
