+++ lib/utils.js
@@ -1,3 +1,4 @@
+import B from 'bluebird';
 import { fs } from 'appium-support';
 import path from 'path';
 import { utils as iosUtils } from 'appium-ios-driver';
@@ -170,7 +171,41 @@ function normalizeCommandTimeouts (value) {
   return result;
 }
 
+async function getPidUsingAppName (udid, appName) {
+  const args = ['-nif', `${appName}.*${udid}`];
+  try {
+    const {stdout} = await exec('pgrep', args);
+    const pid = parseInt(stdout, 10);
+    if (isNaN(pid)) {
+      log.debug(`Cannot parse process id from 'pgrep ${args.join(' ')}' output: ${stdout}`);
+      return null;
+    }
+    return `${pid}`;
+  } catch (err) {
+    log.debug(`'pgrep ${args.join(' ')}' didn't detect any matching processes. Return code: ${err.code}`);
+    return null;
+  }
+}
+
+async function killAppUsingAppName (udid, appName) {
+  const execPkill = async (signal) => {
+    const args = [`-${signal}`, '-if', `${appName}.*${udid}`];
+    try {
+      await exec('pkill', args);
+    } catch (err) {
+      log.debug(`pkill ${args.join(' ')} -> ${err.message}`);
+    }
+  };
+  for (const signal of [2, 15, 9]) {
+    if (!await getPidUsingAppName(udid, appName)) {
+      return;
+    }
+    await execPkill(signal);
+    await B.delay(100);
+  }
+}
+
 export { detectUdid, getAndCheckXcodeVersion, getAndCheckIosSdkVersion,
          adjustWDAAttachmentsPermissions, checkAppPresent, getDriverInfo,
          clearSystemFiles, translateDeviceName, normalizeCommandTimeouts,
-         DEFAULT_TIMEOUT_KEY };
+         DEFAULT_TIMEOUT_KEY, killAppUsingAppName, getPidUsingAppName };
