+++ lib/wda/utils.js
@@ -1,4 +1,3 @@
-import B from 'bluebird';
 import { fs, tempDir } from 'appium-support';
 import { exec } from 'teen_process';
 import path from 'path';
@@ -157,40 +156,6 @@ CODE_SIGN_IDENTITY = ${signingId}
   return xcconfigPath;
 }
 
-async function getPidUsingAppName (udid, appName) {
-  const args = ['-nif', `${appName}.*${udid}`];
-  try {
-    const {stdout} = await exec('pgrep', args);
-    const pid = parseInt(stdout, 10);
-    if (isNaN(pid)) {
-      log.debug(`Cannot parse process id from 'pgrep ${args.join(' ')}' output: ${stdout}`);
-      return null;
-    }
-    return `${pid}`;
-  } catch (err) {
-    log.debug(`'pgrep ${args.join(' ')}' didn't detect any matching processes. Return code: ${err.code}`);
-    return null;
-  }
-}
-
-async function killAppUsingAppName (udid, appName) {
-  const execPkill = async (signal) => {
-    const args = [`-${signal}`, '-if', `${appName}.*${udid}`];
-    try {
-      await exec('pkill', args);
-    } catch (err) {
-      log.debug(`pkill ${args.join(' ')} -> ${err.message}`);
-    }
-  };
-  for (const signal of [2, 15, 9]) {
-    if (!await getPidUsingAppName(udid, appName)) {
-      return;
-    }
-    await execPkill(signal);
-    await B.delay(100);
-  }
-}
-
 async function killProcess (name, proc) {
   if (proc && proc.proc) {
     log.info(`Shutting down ${name} process (pid ${proc.proc.pid})`);
@@ -225,5 +190,5 @@ function randomInt (low, high) {
 }
 
 export { updateProjectFile, resetProjectFile, checkForDependencies,
-         setRealDeviceSecurity, fixForXcode7, fixForXcode9, getPidUsingAppName,
-         generateXcodeConfigFile, killAppUsingAppName, killProcess, randomInt };
+         setRealDeviceSecurity, fixForXcode7, fixForXcode9,
+         generateXcodeConfigFile, killProcess, randomInt };
