+++ lib/wda/webdriveragent.js
@@ -5,7 +5,8 @@ import { JWProxy } from 'appium-base-driver';
 import { fs } from 'appium-support';
 import log from '../logger';
 import { NoSessionProxy } from "./no-session-proxy";
-import { checkForDependencies, killAppUsingAppName } from './utils';
+import { checkForDependencies } from './utils';
+import { killAppUsingAppName } from '../utils';
 import XcodeBuild from './xcodebuild';
 import iProxy from './iproxy';
 import AsyncLock from 'async-lock';
