+++ lib/wda/xcodebuild.js
@@ -5,7 +5,8 @@ import { fs, logger } from 'appium-support';
 import log from '../logger';
 import B from 'bluebird';
 import { fixForXcode7, fixForXcode9, setRealDeviceSecurity, generateXcodeConfigFile,
-         updateProjectFile, resetProjectFile, killProcess, getPidUsingAppName } from './utils';
+         updateProjectFile, resetProjectFile, killProcess } from './utils';
+import { getPidUsingAppName } from '../utils';
 import _ from 'lodash';
 import path from 'path';
